// 1-front/src/app/_services/generated-api/model/listeDeroulanteDTO.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ListeDeroulanteDTO { 
    id: number;
    champ: string;
    valeur: string;
    actif: boolean;
}